Uploaded C program files
In the C Program, we work with a Sentinel Value or Flag value: Flag. A While loop is used on this value and while this Alert value is maintained, the program will continue in execution. In order to resolve this situation without altering the Flag value, it is established that when the counter variable is greater than the value 10, the Break statement is used, which allows breaking the cycle and displaying the result in integer format on the screen.